feat: Update pipelines for saving data and enable in settings
The commit modifies the UdataTestPipeline class to add more 
functionality for processing items. It extracts and saves item data into 
a json file after parsing the URL. The directory for storing json files 
is created if it does not exist. Furthermore, the updated pipeline has 
been enabled in the settings.py file.